refactor(dashboard): extract Mastio policies sub-router (F-B-201 PR-6/10)
Move the three /proxy/policies routes (editor page, save handler with
rules/pdp tab handling, HTMX webhook connectivity probe) out of the
5106-LOC ``mcp_proxy/dashboard/router.py`` god-object into a dedicated
``policies_routes.py`` module, mounted via ``include_router``.

Same pattern as PR-2..PR-5. The sub-router declares paths relative to
/proxy (the prefix inherits from the parent). CSRF gate, approval-hook
intercept, SSRF guard on the test-webhook endpoint and the policy_rules
JSON re-render on parse error all preserved verbatim.

Drive-by cleanup of a stale ``return RedirectResponse`` orphaned in
PR-5 (was unreachable module-level code below the comment marker).

Router footprint: 3592 -> 3454 LOC inline, 66 routes total (3
policies routes now under the new sub-router).

Closes F-B-201 progress: 6/10 PR.
